refactor: update imports to use .jsx extension and convert AppWrapper to arrow function
