+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-03-17T13:24:26.070Z\n"
-"PO-Revision-Date: 2025-03-17T13:24:26.070Z\n"
+"POT-Creation-Date: 2025-03-20T10:33:06.460Z\n"
+"PO-Revision-Date: 2025-03-20T10:33:06.461Z\n"
 
 msgid "Not authorized"
 msgstr "Not authorized"
@@ -516,8 +516,8 @@ msgstr ""
 msgid "Learn more about data entry"
 msgstr "Learn more about data entry"
 
-msgid "Filter data sets in all sections"
-msgstr "Filter data sets in all sections"
+msgid "Filter data elements in all sections"
+msgstr "Filter data elements in all sections"
 
 msgid "Filter fields"
 msgstr "Filter fields"
@@ -594,11 +594,6 @@ msgstr "Type here to filter rows in this section"
 msgid "1 item hidden by filter"
 msgstr "1 item hidden by filter"
 
-msgid "{{count}} items hidden by filter"
-msgid_plural "{{count}} items hidden by filter"
-msgstr[0] "{{count}} items hidden by filter"
-msgstr[1] "{{count}} items hidden by filter"
-
 msgid "{{pendingMutations}} change saved locally"
 msgstr "{{pendingMutations}} change saved locally"
 
@@ -633,3 +628,11 @@ msgstr "Close details sidebar"
 
 msgid "Integer numbers have to be in the range from -2147483648 to 2147483647"
 msgstr "Integer numbers have to be in the range from -2147483648 to 2147483647"
+
+msgctxt "Application title"
+msgid "__MANIFEST_APP_TITLE"
+msgstr "Data Entry"
+
+msgctxt "Application description"
+msgid "__MANIFEST_APP_DESCRIPTION"
+msgstr ""
