+++ src/data-workspace/category-combo-table-body-pivoted/category-combo-table-body-pivoted.jsx
@@ -4,8 +4,8 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { useMetadata, selectors } from '../../shared/index.js'
 import { DataEntryCell, DataEntryField } from '../data-entry-cell/index.js'
-import { getFieldId } from '../get-field-id.js'
-import { TableBodyHiddenByFiltersRow } from '../table-body-hidden-by-filter-row.js'
+import { getFieldId } from '../get-field-id.jsx'
+import { TableBodyHiddenByFiltersRow } from '../table-body-hidden-by-filter-row.jsx'
 import styles from '../table-body.module.css'
 import { generateFormMatrix } from './generate-form-matrix/index.js'
