+++ src/data-workspace/category-combo-table-body/category-combo-table-body.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import React, { useCallback } from 'react'
 import { selectors, useMetadata } from '../../shared/index.js'
 import { DataEntryCell, DataEntryField } from '../data-entry-cell/index.js'
-import { getFieldId } from '../get-field-id.js'
+import { getFieldId } from '../get-field-id.jsx'
 import { TableBodyHiddenByFiltersRow } from '../table-body-hidden-by-filter-row.jsx'
 import styles from '../table-body.module.css'
 import { CategoryComboTableBodyHeader } from './category-combo-table-body-header.jsx'
