+++ src/data-workspace/section-form/section.jsx
@@ -15,7 +15,7 @@ import React, { useMemo, useState } from 'react'
 import { useMetadata, selectors } from '../../shared/index.js'
 import { CategoryComboTableBody } from '../category-combo-table-body/index.js'
 import { PivotedCategoryComboTableBody } from '../category-combo-table-body-pivoted/index.js'
-import { getFieldId } from '../get-field-id.js'
+import { getFieldId } from '../get-field-id.jsx'
 import { IndicatorsTableBody } from '../indicators-table-body/indicators-table-body.jsx'
 import { getDisplayOptions } from './displayOptions.js'
 import { SanitizedText } from './sanitized-text.jsx'
